# services/proj_wesbank/backend/src/app/api/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Wesbank API",
    "description": "Your Site API ⚡",
    "termsOfService": "https://www.YourSite.co.za/termsandconditions.html",
    "contact": { "url": "https://www.YourSite.co.za/#contact" },
    "version": "1.5.0"
  },
  "paths": {
...
    "/v0/get_pav_usage_summary": {
      "get": {
        "tags": ["usage"],
        "summary": "Get Pav Usage Summary",
        "description": "Gets usage summary for a single vehicle",
        "operationId": "get_pav_usage_summary_v0_get_pav_usage_summary_get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Vehiclereg", "type": "string" },
            "name": "vehiclereg",
            "in": "query"
          },
          {
            "required": true,
            "schema": { "title": "Period", "type": "string" },
            "name": "period",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      },
      "post": {
        "tags": ["per assets view"],
        "summary": "Get Pav Usage Summary",
        "description": "Gets usage summary for a single vehicle",
        "operationId": "get_pav_usage_summary_v0_get_pav_usage_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_one_month_cpk_usage_graph": {
      "post": {
        "tags": ["usage"],
        "summary": "Sho002 Get One Month Cpk Usage Graph",
        "description": "Get distance, costs and cpk for a given month for the usage graph",
        "operationId": "sho002_get_one_month_cpk_usage_graph_v0_sho002_get_one_month_cpk_usage_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_all_time_cpk_usage_graph": {
      "post": {
        "tags": ["usage"],
        "summary": "Sho002 Get All Time Cpk Usage Graph",
        "description": "Get distance, costs and cpk for all time per month for the usage graph",
        "operationId": "sho002_get_all_time_cpk_usage_graph_v0_sho002_get_all_time_cpk_usage_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_veh_type_usage_table": {
      "post": {
        "tags": ["usage"],
        "summary": "Sho002 Get Veh Type Usage Table",
        "description": "Get distance, costs and cpk for a given month for the vehicle-type usage table",
        "operationId": "sho002_get_veh_type_usage_table_v0_sho002_get_veh_type_usage_table_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_veh_usage_table": {
      "post": {
        "tags": ["usage"],
        "summary": "Sho002 Get Veh Usage Table",
        "description": "Get distance, costs and cpk for a given month for the vehicle specific usage table",
        "operationId": "sho002_get_veh_usage_table_v0_sho002_get_veh_usage_table_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "token" } }
      }
    }
  }
}